perf(json): use 'vim.json' for nvim-0.9+
